Code reorganization & cleanup
